Suppress Unity websocket exception on graceful socket closure (#220)
It appears that in Unity, calling CloseAsync() can cause a race
condition that sometimes causes `CloseOutputAsync()` to be called
once the underlying socket object has already been disposed.
This exception can be safely suppressed as the method shouldn't
be called on Client initiated socket closure, but only
as a response from a Close message of server initiated closure.

---------

Co-authored-by: Chris Molozian <chris@heroiclabs.com>